requirements.txt
thefuzz>=0.22.1
python-Levenshtein>=0.25.0
requests>=2.31.0
pyyaml>=6.0.1
jinja2>=3.1.3
